requirements.txt
fastapi
uvicorn
groq
pydantic
sentence-transformers
faiss-cpu
numpy
python-multipart
jinja2


